[Frontend] Consolidate scale out entrypoints (#44512)
