+++ docs/serving/online_serving/README.md
@@ -119,9 +119,9 @@ For further details on profiling vLLM, please refer to [this page](../../contrib
 - `/ping` - SageMaker health check
 - `/invocations` - SageMaker-compatible endpoint (routes to the same inference functions as `/v1` endpoints)
 
-## Disaggregated Everything
+## Scale-Out APIs
 
-### Tokens IN <> Tokens OUT
+### Tokens IN <> Tokens OUT APIs
 
 - `/inference/v1/generate` - Generate completions
 - `/abort_requests` - Abort in-flight requests (only when `--tokens-only` is also set)
