+++ examples/scale_out/token_generation_client.py
