+++ tests/entrypoints/openai/test_render_token_offsets.py
@@ -3,7 +3,7 @@
 """Unit tests for the token-offsets request/response protocol wiring:
 the request flag flowing into ``TokenizeParams`` and the ``GenerateRequest``
 serialization boundary. End-to-end behavior is covered by
-``tests/entrypoints/serve/render/test_render.py``; plain Pydantic field
+``tests/entrypoints/scale_out/render/test_render.py``; plain Pydantic field
 storage is not retested here.
 """
 
@@ -12,7 +12,7 @@ from unittest.mock import Mock
 from vllm.config import ModelConfig
 from vllm.entrypoints.openai.chat_completion.protocol import ChatCompletionRequest
 from vllm.entrypoints.openai.completion.protocol import CompletionRequest
-from vllm.entrypoints.serve.disagg.protocol import GenerateRequest
+from vllm.entrypoints.scale_out.token_in_token_out.protocol import GenerateRequest
 from vllm.sampling_params import SamplingParams
