+++ tests/entrypoints/scale_out/derender/test_derender.py
