+++ tests/entrypoints/scale_out/render/test_launch_render.py
