+++ tests/entrypoints/scale_out/render/test_render.py
