+++ tests/entrypoints/scale_out/render/test_render_multimodal.py
