+++ tests/entrypoints/scale_out/token_in_token_out/test_serving_multimodal_tokens.py
