+++ vllm/entrypoints/scale_out/derender/api_router.py
@@ -1,0 +1,74 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+from http import HTTPStatus
+
+from fastapi import APIRouter, Depends, Request
+from fastapi.responses import JSONResponse
+
+from vllm.entrypoints.openai.chat_completion.protocol import ChatCompletionResponse
+from vllm.entrypoints.openai.completion.protocol import CompletionResponse
+from vllm.entrypoints.openai.engine.protocol import ErrorResponse
+from vllm.entrypoints.serve.utils.api_utils import validate_json_request
+from vllm.logger import init_logger
+
+from ..token_in_token_out.protocol import (
+    DerenderChatRequest,
+    DerenderCompletionRequest,
+)
+from .serving import ServingDerender
+
+logger = init_logger(__name__)
+
+router = APIRouter()
+
+
+def derender(request: Request) -> ServingDerender | None:
+    return getattr(request.app.state, "serving_derender", None)
+
+
+@router.post(
+    "/v1/chat/completions/derender",
+    dependencies=[Depends(validate_json_request)],
+    response_model=ChatCompletionResponse,
+    responses={
+        HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
+        HTTPStatus.NOT_FOUND.value: {"model": ErrorResponse},
+        HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
+    },
+)
+async def derender_chat_completion(request: DerenderChatRequest, raw_request: Request):
+    handler = derender(raw_request)
+    if handler is None:
+        raise NotImplementedError(
+            "The model does not support Chat Completions Derender API"
+        )
+
+    result = await handler.derender_chat_response(request)
+
+    if isinstance(result, ErrorResponse):
+        return JSONResponse(content=result.model_dump(), status_code=result.error.code)
+
+    return JSONResponse(content=result.model_dump())
+
+
+@router.post(
+    "/v1/completions/derender",
+    dependencies=[Depends(validate_json_request)],
+    response_model=CompletionResponse,
+    responses={
+        HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
+        HTTPStatus.NOT_FOUND.value: {"model": ErrorResponse},
+        HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
+    },
+)
+async def derender_completion(request: DerenderCompletionRequest, raw_request: Request):
+    handler = derender(raw_request)
+    if handler is None:
+        raise NotImplementedError("The model does not support Completions Derender API")
+
+    result = await handler.derender_completion_response(request)
+
+    if isinstance(result, ErrorResponse):
+        return JSONResponse(content=result.model_dump(), status_code=result.error.code)
+
+    return JSONResponse(content=result.model_dump())
