+++ vllm/entrypoints/scale_out/render/api_router.py
@@ -2,26 +2,17 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 from http import HTTPStatus
 
-from fastapi import APIRouter, Depends, FastAPI, Request
+from fastapi import APIRouter, Depends, Request
 from fastapi.responses import JSONResponse
 
-from vllm.entrypoints.openai.chat_completion.protocol import (
-    ChatCompletionRequest,
-    ChatCompletionResponse,
-)
-from vllm.entrypoints.openai.completion.protocol import (
-    CompletionRequest,
-    CompletionResponse,
-)
+from vllm.entrypoints.openai.chat_completion.protocol import ChatCompletionRequest
+from vllm.entrypoints.openai.completion.protocol import CompletionRequest
 from vllm.entrypoints.openai.engine.protocol import ErrorResponse
-from vllm.entrypoints.serve.disagg.protocol import (
-    DerenderChatRequest,
-    DerenderCompletionRequest,
-    GenerateRequest,
-)
-from vllm.entrypoints.serve.render.serving import ServingRender
 from vllm.entrypoints.serve.utils.api_utils import validate_json_request
 from vllm.logger import init_logger
+
+from ..token_in_token_out.protocol import GenerateRequest
+from .serving import ServingRender
 
 logger = init_logger(__name__)
 
@@ -79,55 +70,3 @@ async def render_completion(request: CompletionRequest, raw_request: Request):
         return JSONResponse(content=result.model_dump(), status_code=result.error.code)
 
     return JSONResponse(content=[item.model_dump() for item in result])
-
-
-@router.post(
-    "/v1/chat/completions/derender",
-    dependencies=[Depends(validate_json_request)],
-    response_model=ChatCompletionResponse,
-    responses={
-        HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
-        HTTPStatus.NOT_FOUND.value: {"model": ErrorResponse},
-        HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
-    },
-)
-async def derender_chat_completion(request: DerenderChatRequest, raw_request: Request):
-    handler = render(raw_request)
-    if handler is None:
-        raise NotImplementedError(
-            "The model does not support Chat Completions Derender API"
-        )
-
-    result = await handler.derender_chat_response(request)
-
-    if isinstance(result, ErrorResponse):
-        return JSONResponse(content=result.model_dump(), status_code=result.error.code)
-
-    return JSONResponse(content=result.model_dump())
-
-
-@router.post(
-    "/v1/completions/derender",
-    dependencies=[Depends(validate_json_request)],
-    response_model=CompletionResponse,
-    responses={
-        HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
-        HTTPStatus.NOT_FOUND.value: {"model": ErrorResponse},
-        HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
-    },
-)
-async def derender_completion(request: DerenderCompletionRequest, raw_request: Request):
-    handler = render(raw_request)
-    if handler is None:
-        raise NotImplementedError("The model does not support Completions Derender API")
-
-    result = await handler.derender_completion_response(request)
-
-    if isinstance(result, ErrorResponse):
-        return JSONResponse(content=result.model_dump(), status_code=result.error.code)
-
-    return JSONResponse(content=result.model_dump())
-
-
-def attach_router(app: FastAPI) -> None:
-    app.include_router(router)
